Manager/requirements.txt
aioredis==1.3.1
amqp==2.6.1
argon2-cffi==20.1.0
asgiref==3.3.1
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.7
bcrypt==4.0.1
billiard==3.6.4.0
celery==4.4.7
certifi==2021.5.30
cffi==1.15.1
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
cryptography==3.4.6
daphne==3.0.1
defusedxml==0.6.0
dj-static==0.0.6
Django==2.2.5
django-braces==1.14.0
django-cors-headers==3.7.0
django-cors-middleware==1.5.0
django-countries==7.0
django-crispy-forms==1.9.2
django-debug-toolbar==3.2
django-filter==2.4.0
django-jet==1.0.8
djangorestframework==3.12.2
djangorestframework-simplejwt==5.0.0
Faker==4.1.0
flower==0.9.7
gunicorn==20.1.0
hiredis==2.0.0
humanize==3.7.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
isort==5.6.4
kombu==4.6.11
lazy-object-proxy==1.4.3
Markdown==3.3.3
mccabe==0.6.1
msgpack==1.0.2
oauthlib==3.1.0
Pillow==9.3.0
pipdeptree==2.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.10.1
PyJWT==2.0.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2019.3
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.0
static3==0.7.0
text-unidecode==1.3
toml==0.10.2
tornado==6.1
Twisted==21.2.0
#twisted-iocpsupport==1.0.1
txaio==21.2.1
urllib3==1.26.3
vine==1.3.0
wcwidth==0.2.5
wincertstore==0.2
wrapt==1.12.1
zope.interface==5.2.0
